+++ service/flex/v1/plugin_configuration/plugins/api_op_client.go
@@ -1,0 +1,25 @@
+// Package plugins contains auto-generated files. DO NOT MODIFY
+package plugins
+
+import "github.com/RJPearson94/twilio-sdk-go/client"
+
+// Client for managing plugin resources
+type Client struct {
+	client *client.Client
+
+	configurationSid string
+}
+
+// ClientProperties are the properties required to manage the plugins resources
+type ClientProperties struct {
+	ConfigurationSid string
+}
+
+// New creates a new instance of the plugins client
+func New(client *client.Client, properties ClientProperties) *Client {
+	return &Client{
+		client: client,
+
+		configurationSid: properties.ConfigurationSid,
+	}
+}