+++ CHANGELOG.md
@@ -21,6 +21,7 @@ FEATURES
 - **New Resource:** Plugin
 - **New Resource:** Plugin Version
 - **New Resource:** Plugin Configuration
+- **New Resource:** Configuration Plugins
 
 ### Sync